Deprecate `ontolius::hierarchy` and `ontolius::prelude`.
